nav-bar and a popup is added to the map.
A navigation bar with a "Select Location" button is made and on click there will be point vector on the map. After putting a point on a map a jsPanel is popped-up with the readonly coordinations with an input for the user to type in the name for that point on the map and an "Add Location" button.