POST and GET http attributes are added.
using fetch post and get http attributes are connected with the map. So, now we can look and see all the previously saved locations from the database on the map. I also added a warning when the user wants to add a location but the user has not zoomed on the map as much as needed.